Create data dir if it doesn't exist
Pusher used to handled creating data directories if they didn't already
exist. We removed that functionality from Pusher because it seems more
appropriate for the producer of a datatype to create its own data
directory. However, we found that pcap did not do this. This commit
should cause pcap to create its own data directory.